0715SwingAxe

+++ HungryJimmy/Assets/01.Scripts/StatusController.cs
@@ -46,10 +46,6 @@ public class StatusController : MonoBehaviour
 
     public bool isRain = false; // 비가 오는지 확인
 
-    // public bool isRunning = false; //run 상태인지 확인
-    // public GameObject runningImage; //run 상태일 때 활성화되는 이미지
-
-
     private const int HUNGRY = 0, THIRSTY = 1, STAMINA = 2;
 
     void Start()
@@ -121,11 +117,7 @@ public class StatusController : MonoBehaviour
                 currentStamina--;
                 currentStaminaDecreaseTime = 0;
             }
-            // if(!isRunning)
-            // {
-            //     Running();
-            //     isRunning = false;
-            // }
+
         }
         else        // 0보다 작아졌을때
             theHealth.Dead();
@@ -139,14 +131,6 @@ public class StatusController : MonoBehaviour
             currentThirsty += 3f * Time.deltaTime; //총 30의 수분 주기(비가 10초동안 옴)
         }
     }
-
-    // private void Running() //run 상태
-    // {
-    //     if (runningImage.activeInHierarchy)
-    //     {
-    //         currentStamina -= 2f * Time.deltaTime; 
-    //     }
-    // }
 
     private void GaugeUpdate()      // 상태 수치 변화 시각화
     {
@@ -167,7 +151,7 @@ public class StatusController : MonoBehaviour
     }
 
     // Stamina 감소
-    public void DecreaseStamina(int _count)
+    public void DecreaseStamina(float _count)
     {
         currentStamina -= _count;
